Moved OutputOptions class to util.outputoptions
